+++ src/scene/shader-lib/chunks/gsplat/vert/gsplatUnpack.js
@@ -1,0 +1,51 @@
+export default /* glsl */`
+vec3 unpack111011(uint bits) {
+    return vec3(
+        float(bits >> 21u) / 2047.0,
+        float((bits >> 11u) & 0x3ffu) / 1023.0,
+        float(bits & 0x7ffu) / 2047.0
+    );
+}
+
+vec4 unpack8888(uint bits) {
+    return vec4(
+        float(bits >> 24u) / 255.0,
+        float((bits >> 16u) & 0xffu) / 255.0,
+        float((bits >> 8u) & 0xffu) / 255.0,
+        float(bits & 0xffu) / 255.0
+    );
+}
+
+const float norm = 1.0 / (sqrt(2.0) * 0.5);
+
+vec4 unpackRotation(uint bits) {
+    float a = (float((bits >> 20u) & 0x3ffu) / 1023.0 - 0.5) * norm;
+    float b = (float((bits >> 10u) & 0x3ffu) / 1023.0 - 0.5) * norm;
+    float c = (float(bits & 0x3ffu) / 1023.0 - 0.5) * norm;
+    float m = sqrt(1.0 - (a * a + b * b + c * c));
+
+    uint mode = bits >> 30u;
+    if (mode == 0u) return vec4(m, a, b, c);
+    if (mode == 1u) return vec4(a, m, b, c);
+    if (mode == 2u) return vec4(a, b, m, c);
+    return vec4(a, b, c, m);
+}
+
+mat3 quatToMat3(vec4 R) {
+    float x = R.x;
+    float y = R.y;
+    float z = R.z;
+    float w = R.w;
+    return mat3(
+        1.0 - 2.0 * (z * z + w * w),
+              2.0 * (y * z + x * w),
+              2.0 * (y * w - x * z),
+              2.0 * (y * z - x * w),
+        1.0 - 2.0 * (y * y + w * w),
+              2.0 * (z * w + x * y),
+              2.0 * (y * w + x * z),
+              2.0 * (z * w - x * y),
+        1.0 - 2.0 * (y * y + z * z)
+    );
+}
+`;
